requirements.txt
requests
pandas
python-dotenv
pymysql
streamlit
plotly-express
pytest
pytest-cov
pytest-mock